Move extractors and factory into its own package
