+++ src/main/java/edu/upf/taln/dri/lib/model/util/extractor/Extractor.java
@@ -1,4 +1,4 @@
-package edu.upf.taln.dri.lib.model.util;
+package edu.upf.taln.dri.lib.model.util.extractor;
 
 import edu.upf.taln.dri.lib.model.DocCacheManager;
 import edu.upf.taln.dri.lib.model.ext.Author;
